# Exclude events table from using window function for pagination