--- a/spec/unit/lib/cloud_controller/paging/sequel_paginator_spec.rb
+++ b/spec/unit/lib/cloud_controller/paging/sequel_paginator_spec.rb
@@ -129,14 +129,14 @@ module VCAP::CloudController
         expect(paginated_result.total).to be > 1
       end
 
-      context 'calls DB twice if table is in exclude-from-window-function-with-pagination list' do
-        let(:dataset) { VCAP::CloudController.const_get(VCAP::CloudController::Event.to_s.classify).dataset }
-        let!(:event_1) { VCAP::CloudController.const_get(VCAP::CloudController::Event.to_s.classify).make(guid: '1', created_at: '2022-12-20T10:47:01Z') }
-        let!(:event_2) { VCAP::CloudController.const_get(VCAP::CloudController::Event.to_s.classify).make(guid: '2', created_at: '2022-12-20T10:47:02Z') }
-        let!(:event_3) { VCAP::CloudController.const_get(VCAP::CloudController::Event.to_s.classify).make(guid: '3', created_at: '2022-12-20T10:47:03Z') }
-        let!(:event_4) { VCAP::CloudController.const_get(VCAP::CloudController::Event.to_s.classify).make(guid: '4', created_at: '2022-12-20T10:47:04Z') }
+      context 'events table' do
+        let(:dataset) { Event.dataset }
+        let!(:event_1) { Event.make(guid: '1', created_at: '2022-12-20T10:47:01Z') }
+        let!(:event_2) { Event.make(guid: '2', created_at: '2022-12-20T10:47:02Z') }
+        let!(:event_3) { Event.make(guid: '3', created_at: '2022-12-20T10:47:03Z') }
+        let!(:event_4) { Event.make(guid: '4', created_at: '2022-12-20T10:47:04Z') }
 
-        it 'does not use window function for events table' do
+        it 'does not use window function' do
           options = { page: page, per_page: per_page }
           pagination_options = PaginationOptions.new(options)