The events table is one of the biggest tables. We saw slow queries on big landscapes like:
```
cf curl "/v3/audit_events?types=audit.user.organization_user_add%2Caudit.user.organization_billing_manager_remove%2Caudit.user.space_developer_add%2Caudit.user.organization_auditor_remove%2Caudit.user.space_developer_remove%2Caudit.user.space_auditor_remove%2Caudit.user.space_auditor_add%2Caudit.user.space_manager_add%2Caudit.user.organization_billing_manager_add%2Caudit.user.organization_manager_remove%2Caudit.user.organization_auditor_add%2Caudit.user.space_manager_remove%2Caudit.user.organization_manager_add%2Caudit.user.organization_user_remove&per_page=500&order_by=-created_at" 
```
We experimented and discovered using v2 for audit_events is much faster. The difference between v2 and v3 is that for v3 pagination is done via window function. 
```
SELECT *, count(*) OVER () AS "pagination_total_results" FROM "events" WHERE ("type" IN ('audit.user.organization_user_add', 'audit.user.organization_billing_manager_remove', 'audit.user.space_developer_add', 'audit.user.organization_auditor_remove', 'audit.user.space_developer_remove', 'audit.user.space_auditor_remove', 'audit.user.space_auditor_add', 'audit.user.space_manager_add', 'audit.user.organization_billing_manager_add', 'audit.user.organization_manager_remove', 'audit.user.organization_auditor_add', 'audit.user.space_manager_remove', 'audit.user.organization_manager_add', 'audit.user.organization_user_remove')) ORDER BY "events"."created_at" DESC, "events"."guid" ASC LIMIT 500 OFFSET 0
```
In v2 it was done with 2 queries.
```
SELECT * FROM "events" WHERE (type  IN  ('audit.user.organization_user_add', 'audit.user.organization_billing_manager_remove', 'audit.user.space_developer_add', 'audit.user.organization_auditor_remove', 'audit.user.space_developer_remove', 'audit.user.space_auditor_remove', 'audit.user.space_auditor_add', 'audit.user.space_manager_add', 'audit.user.organization_billing_manager_add', 'audit.user.organization_manager_remove', 'audit.user.organization_auditor_add', 'audit.user.space_manager_remove', 'audit.user.organization_manager_add', 'audit.user.organization_user_remove')) ORDER BY "timestamp" DESC, "id" DESC LIMIT 500 OFFSET 0

SELECT count(*) AS "count" FROM "events" WHERE (type  IN  ('audit.user.organization_user_add', 'audit.user.organization_billing_manager_remove', 'audit.user.space_developer_add', 'audit.user.organization_auditor_remove', 'audit.user.space_developer_remove', 'audit.user.space_auditor_remove', 'audit.user.space_auditor_add', 'audit.user.space_manager_add', 'audit.user.organization_billing_manager_add', 'audit.user.organization_manager_remove', 'audit.user.organization_auditor_add', 'audit.user.space_manager_remove', 'audit.user.organization_manager_add', 'audit.user.organization_user_remove')) LIMIT 1
```
For big tables like the events table using pagination with window function seems not to be the best choice. 
With this PR we change it so that the events table is excluded from window function for pagiantion
The query for the /v3/audit_events call would look like following with current change:
(no more window function)
```
SELECT * FROM "events" WHERE (type  IN  ('audit.user.organization_user_add', 'audit.user.organization_billing_manager_remove', 'audit.user.space_developer_add', 'audit.user.organization_auditor_remove', 'audit.user.space_developer_remove', 'audit.user.space_auditor_remove', 'audit.user.space_auditor_add', 'audit.user.space_manager_add', 'audit.user.organization_billing_manager_add', 'audit.user.organization_manager_remove', 'audit.user.organization_auditor_add', 'audit.user.space_manager_remove', 'audit.user.organization_manager_add', 'audit.user.organization_user_remove')) ORDER BY "events"."created_at" DESC, "events"."guid" ASC LIMIT 500 OFFSET 0

SELECT count(*) AS "count" FROM "events" WHERE (type  IN  ('audit.user.organization_user_add', 'audit.user.organization_billing_manager_remove', 'audit.user.space_developer_add', 'audit.user.organization_auditor_remove', 'audit.user.space_developer_remove', 'audit.user.space_auditor_remove', 'audit.user.space_auditor_add', 'audit.user.space_manager_add', 'audit.user.organization_billing_manager_add', 'audit.user.organization_manager_remove', 'audit.user.organization_auditor_add', 'audit.user.space_manager_remove', 'audit.user.organization_manager_add', 'audit.user.organization_user_remove')) LIMIT 1
```
Excluding the events table from using the pagination with window function and do 2 queries like it was in v2 improves query time extremely.

|cf curl /v3/  on 12813074 events|time in s with window function |time in s without window function |
| ----- | ----- | ----- |
|'/v3/audit_events'	|30,34|1,94|
|'/v3/audit_events?page=1&per_page=5&created_ats[gt]=2022-11-23T15:36:54Z'|	22,10|2,17|
|'/v3/audit_events?types=...&per_page=100&order_by=-created_at'|	1,48|	0,69|
|'/v3/audit_events?space_guids=dd3a8c27-bc6f-4b58-930d-d77865c9a428&page=1&per_page=5&order_by=-created_at'	|54,41|	8,51|

|cf curl /v3/  on 1413076 events|time in s with window function |time in s without window function |
| ----- | ----- | ----- |
|'/v3/audit_events'	|2.324|0.198	|
|'/v3/audit_events?page=1&per_page=5&created_ats[gt]=2022-11-23T15:36:54Z'|	0.651|	0.207|
|'/v3/audit_events?types=...&per_page=100&order_by=-created_at'|	1,48|0.338|
|'/v3/audit_events?space_guids=dd3a8c27-bc6f-4b58-930d-d77865c9a428&page=1&per_page=5&order_by=-created_at'	|4.542|	1.483|

![query-time-v2-vs-v3-audit-events-with-type-filter](https://user-images.githubusercontent.com/30441792/208475439-914563ca-b49a-44ed-9d9f-e5d58f26b4f7.png)
Call that were used for graph creation (for v2 of course equivalent v2 call):
```
cf curl "/v3/audit_events?types=audit.user.organization_user_add%2Caudit.user.organization_billing_manager_remove%2Caudit.user.space_developer_add%2Caudit.user.organization_auditor_remove%2Caudit.user.space_developer_remove%2Caudit.user.space_auditor_remove%2Caudit.user.space_auditor_add%2Caudit.user.space_manager_add%2Caudit.user.organization_billing_manager_add%2Caudit.user.organization_manager_remove%2Caudit.user.organization_auditor_add%2Caudit.user.space_manager_remove%2Caudit.user.organization_manager_add%2Caudit.user.organization_user_remove&per_page=100&order_by=-created_at"
```


Thanks for contributing to cloud_controller_ng. To speed up the process of reviewing your pull request please provide us with:

* A short explanation of the proposed change:

* An explanation of the use cases your change solves

* Links to any other associated PRs

* [X] I have reviewed the [contributing guide](https://github.com/cloudfoundry/cloud_controller_ng/blob/main/CONTRIBUTING.md)

* [X] I have viewed, signed, and submitted the Contributor License Agreement

* [X] I have made this pull request to the `main` branch

* [X] I have run all the unit tests using `bundle exec rake`

* [ ] I have run [CF Acceptance Tests](https://github.com/cloudfoundry/cloud_controller_ng/blob/main/spec/README.md#cf-acceptance-tests-cats)